scripts/requirements.txt
python-dotenv
playwright
beautifulsoup4
discord.py
requests
